+++ collections/src/main/kotlin/ktx/collections/arrays.kt
@@ -237,7 +237,7 @@ inline fun <Type, R : Comparable<R>> GdxArray<out Type>.sortByDescending(crossin
  * Removes elements from the array that satisfy the [predicate].
  * @param pool Removed items are freed to this pool.
  */
-inline fun <Type> GdxArray<Type>.removeAll(pool: Pool<Type>?, predicate: (Type) -> Boolean) {
+inline fun <Type> GdxArray<Type>.removeAll(pool: Pool<Type>? = null, predicate: (Type) -> Boolean) {
   var currentWriteIndex = 0
   for (i in 0 until size) {
     val value = items[i]
@@ -268,6 +268,26 @@ inline fun <Type> GdxArray<Type>.retainAll(pool: Pool<Type>? = null, predicate: 
       currentWriteIndex++
     } else {
       pool?.free(value)
+    }
+  }
+  truncate(currentWriteIndex)
+}
+
+/**
+ * Transfers elements that match the [predicate] into the selected [toArray].
+ * The elements will be removed from this array and added [toArray].
+ */
+inline fun <Type> GdxArray<Type>.transfer(toArray: GdxArray<Type>, predicate: (Type) -> Boolean) {
+  var currentWriteIndex = 0
+  for (i in 0 until size) {
+    val value = items[i]
+    if (predicate(value)) {
+      toArray.add(value)
+    } else {
+      if (currentWriteIndex != i) {
+        items[currentWriteIndex] = value
+      }
+      currentWriteIndex++
     }
   }
   truncate(currentWriteIndex)