+++ collections/src/test/kotlin/ktx/collections/arraysTest.kt
@@ -394,6 +394,58 @@ class ArraysTest {
   }
 
   @Test
+  fun `should transfer elements matching predicate`() {
+    val array = GdxArray.with(0, 1, 2, 3, 4)
+    val target = GdxArray.with<Int>()
+
+    array.transfer(toArray = target) {
+      it % 2 == 0
+    }
+
+    assertEquals(GdxArray.with(1, 3), array)
+    assertEquals(GdxArray.with(0, 2, 4), target)
+  }
+
+  @Test
+  fun `should transfer all elements`() {
+    val array = GdxArray.with(0, 1, 2, 3, 4)
+    val target = GdxArray.with<Int>()
+
+    array.transfer(toArray = target) {
+      true
+    }
+
+    assertEquals(GdxArray.with<Int>(), array)
+    assertEquals(GdxArray.with(0, 1, 2, 3, 4), target)
+  }
+
+  @Test
+  fun `should transfer no elements`() {
+    val array = GdxArray.with(0, 1, 2, 3, 4)
+    val target = GdxArray.with<Int>()
+
+    array.transfer(toArray = target) {
+      false
+    }
+
+    assertEquals(GdxArray.with(0, 1, 2, 3, 4), array)
+    assertEquals(GdxArray.with<Int>(), target)
+  }
+
+  @Test
+  fun `should not transfer any elements from empty array`() {
+    val array = GdxArray.with<Int>()
+    val target = GdxArray.with<Int>()
+
+    array.transfer(toArray = target) {
+      it % 2 == 0
+    }
+
+    assertEquals(GdxArray.with<Int>(), array)
+    assertEquals(GdxArray.with<Int>(), target)
+  }
+
+  @Test
   fun `should retain elements from existing GdxArray`() {
     val array = GdxArray.with(1, 2, 3, 4, 5)
     array.retainAll { it < 6 }