+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - **[UPDATE]** Updated to Kotlin 1.3.61.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.3.3.
 - **[FEATURE]** (`ktx-assets`) Added `AssetGroup` abstract class that allows to manage groups of assets.
-- **[FEATURE]** (`ktx-collections`) Added `Array.removeAll` and `Array.retainAll` higher-order in-place filters.
+- **[FEATURE]** (`ktx-collections`) Added `removeAll`, `retainAll` and `transfer` extensions to LibGDX `Array` using lambda predicates to modify the array in-place.
 - **[CHANGE]** (`ktx-collections`) `PooledList` now implements `MutableIterable`.
 - **[FEATURE]** (`ktx-graphics`) Added `Batch.begin` extension methods that automatically set projection matrix from a `Camera` or `Matrix4`. 
 - **[FEATURE]** (`ktx-style`) Added `Skin` extension methods with reified resource types: `optional`, `add`, `remove`, `has` and `getAll`.